LayoutTests/imported/w3c/web-platform-tests/css/css-contain/container-queries/container-longhand-animation-type-expected.txt


FAIL Reference variable is applied assert_equals: expected "PASS" but got "FAIL"
PASS container-name is not animatable
PASS container-type is not animatable
